--refactor: email service
