+++ config/index.js
@@ -5,6 +5,8 @@ module.exports = {
 	PORT: process.env.PORT,
 	EMAIL_USER: process.env.EMAIL_USER,
 	EMAIL_PASS: process.env.EMAIL_PASS,
+	EMAIL_HOST: process.env.EMAIL_HOST,
+	EMAIL_SENDER: process.env.EMAIL_SENDER,
 	GET_BIRTHDAYS_PROCESS_API: process.env.GET_BIRTHDAYS_PROCESS_API,
 	GET_BIRTHDAYS_PROCESS_SECRET: process.env.GET_BIRTHDAYS_PROCESS_SECRET,
 	WISH_MESSAGE: "Happy Birthday!!",
