+++ package.json
@@ -16,6 +16,7 @@
         "log4js": "^6.1.2",
         "node-cron": "^2.0.3",
         "nodemailer": "^6.4.6",
+        "nodemailer-html-to-text": "^3.1.0",
         "nodemon": "^1.19.3"
     }
 }
